# docs/swagger/user.yaml
openapi: 3.0.3
info:
  title: Clean Architecture Fiber Starter API
  version: 0.1.0
  description: Sample User domain demonstrating clean architecture layering.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /auth/login:
    get:
      summary: Start OIDC login
      description: Redirects to the configured OpenID Connect provider using Authorization Code with PKCE.
      operationId: startOIDCLogin
      responses:
        '302':
          description: Redirect to OIDC provider
        '503':
          description: OIDC disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /auth/callback:
    get:
      summary: Complete OIDC login
      description: Validates the provider callback and returns a local Bearer access token.
      operationId: completeOIDCLogin
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Login completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Invalid callback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Login not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /users:
    post:
      summary: Register user
      description: Creates a new user and returns the public profile.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Users
  /users/{id}:
    get:
      summary: Fetch user by ID
      description: Retrieves a user profile by identifier. Requires authentication.
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the user.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Users
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    AuthTokenResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              format: int64
            user:
              $ref: '#/components/schemas/UserResponse'
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            detail:
              type: string